PhotoAPI: Take sensor orientation into account.
On a Nexus 5X, the old code would take pictures that were upside-down.
Apparently, the Android device standards require that the long edge of the
sensor must align with the long edge of the device, but doesn't say which of the
two ways to satisfy this manufacturers should pick. This change makes PhotoAPI
generate right-side-up pictures in both conditions.

Tested on a Nexus 7 (second-gen) and a Nexus 5X.