MicRecorderAPI: a couple of fixes
Have year before month and day in the default file name for better
sorting. Also fix the extension name in it for automatic format
selection. As we want the file to have extension name by default,
we have to pick a default format/encoder. AAC is chosen because it
is probably the most compatible one among different platforms.