MicRecoderAPI: a few enhancements (#248)
* MicRecorderAPI: allow duration limit to be disabled with zero or negative

* MicRecorderAPI: avoid unnecessary MediaRecorder parameter in RecorderCommandHandler.handle()

* MicRecorderAPI: allow encoder and format to be set by user

All supported encoders can be set with their names. Unless a format
is set explicitly with its corresponding integer, a suitable format
would be chosen automatically for the chosen encoder.

* MicRecorderAPI: allow AudioSource to be set by user

Just in case, so it can only be set with the source's corresponding
integer.

* MicRecorderAPI: allow bit rate, sampling rate and number of channels to be set by user

* MicRecorderAPI: a couple of fixes

Have year before month and day in the default file name for better
sorting. Also fix the extension name in it for automatic format
selection. As we want the file to have extension name by default,
we have to pick a default format/encoder. AAC is chosen because it
is probably the most compatible one among different platforms.

* MicRecorderAPI: replace all getAbsoluteFile() with getAbsolutePath()

None of them were used in places where it should be a File.

* MicRecorderAPI: properly report unlimited MaxDuration

* MicRecorderAPI: make sure the recorder is released when not used

* MicRecorderAPI: remove aac_eld/he_aac/vorbis support

They are at least problematic on my phone. They either fail to work
or crash the API app. Sometimes the crash is even unrecoverable.