Fix for issue #295
Changed sensor pattern matching to first look for a completely matching sensor before looking for a partial match. 

I know it's gross but a else-if wouldn't work due to still breaking out due to partial match before checking all sensors for a complete match. I would love to find out how to do this without a flag variable and duplicating so much code.