Resolved crashing by copying sensor list before sort
getSensorList returns a unmodifiable list causing a UnsupportedOperationException to be thrown on sort crashing the api. This is resolved by changing availableSensors to create a copy of what getSensorList returns rather than assigning what is returns directly to availableSensors.  

Sorting with a lambda rather than  comparator.comparing() maintains the current api level.