Changed: Use fully qualified name for resource variable references of other modules/libraries as required by Android gradle plugin (`AGP`) version `>= 8.0.0`
The default value for `android.nonTransitiveRClass` property has been changed from `false` to `true`. The value can be overridden to `false` in the global/root `gradle.properties` file of the project, but that will affect all modules, so instead we just use fully qualified name in `app` module that is triggering build errors.

- https://developer.android.com/build/releases/past-releases/agp-8-0-0-release-notes#default-changes
- https://developer.android.com/build/optimize-your-build#use-non-transitive-r-classes