Changed: Rename `TermuxAPIActivity` to `TermuxAPIMainActivity` stage 2
The `TermuxAPILauncherActivity` activity alias has been added for `TermuxAPIMainActivity`. This will allow disabling the launcher activity, but still allow the main activity to be launched from Termux app. The launcher activity will also now be allowed to be enabled again without having to reinstall the app with the `Enable Launcher Icon` button in the main activity if its currently disabled.